% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradepolicy-package.R
\docType{data}
\name{agtpa_applications}
\alias{agtpa_applications}
\title{International Trade Data for Application Exercises in AGTPA}
\format{
A data frame object of 99,981 x 9
}
\description{
Contains bilateral trade flows for different pairs of countries
between 1986 and 2006 and additional variables modelling. See the details
in Yotov, et al (2016) \url{isbn:978-92-870-4367-2}.
}
\author{
United Nations Conference on Trade and Development (UNCTAD)
}
\keyword{data}
