% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{tp_clustered_glm}
\alias{tp_clustered_glm}
\title{GLM Regression With Clustered Standard Errors}
\usage{
tp_clustered_glm(
  formula,
  data,
  pair = "pair_id",
  fe_pattern = "^exporter|^importer"
)
}
\arguments{
\item{formula}{A formula for the model}

\item{data}{A tibble or data.frame}

\item{pair}{Inter-national fixed effects column (defaults to "pair_id")}

\item{fe_pattern}{A pattern for the fixed effects variable, allows character or regex (defaults to "^exporter|^importer")}
}
\value{
a tibble
}
\description{
Fits a regression with robust clustered standard errors. This uses a quasi-poisson
family and returns the estimated coefficients after computing a clustered
variance-covariance matrix.
}
