% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{tp_database}
\alias{tp_database}
\title{The local AGTPA database}
\usage{
tp_database(dir = tp_path())
}
\arguments{
\item{dir}{The location of the database on disk. Defaults to
\code{tradepolicy} under \code{tools::R_user_dir("tradepolicy")}, or the path specified
in the environment variable \code{TRADEPOLICY_DIR}.}
}
\description{
Returns a local database connection, which is a DuckDB DBI-compatible connection.
Unlike \code{\link[=tp_table]{tp_table()}}, this function is more flexible and can be used
with dbplyr to read exactly what you need or with DBI to use SQL commands.
}
\examples{
\dontrun{
 DBI::dbListTables(tp_database())

 DBI::dbGetQuery(
  tp_database(),
  "SELECT * FROM ch1_application1 WHERE exporter = 'CAN'"
 )
}
}
