% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{tp_disconnect}
\alias{tp_disconnect}
\title{Disconnect from the AGTPA database}
\usage{
tp_disconnect()
}
\description{
A utility function for disconnecting from the database.
}
\examples{
tp_disconnect()
}
