% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{tp_download}
\alias{tp_download}
\title{Download the Yotov database to your local computer}
\usage{
tp_download(ver = NULL, verbose = interactive())
}
\arguments{
\item{verbose}{Whether to display messages and download progress}

\item{tag}{What release tag of data to download. Defaults to the most recent.
Releases are expected to come twice per year. See all releases at
\url{https://github.com/pachamaltese/tradepolicy/releases}.}

\item{destdir}{Where to download the compressed file.}

\item{cleanup}{Whether to delete the compressed file after loading into the database.}
}
\description{
This command downloads the Yotov trade database and populates a local
database. The download is 31.4 MB, and the database uses 3 GB on disk.
During import over 3.5 GB of disk space may be used temporarily.
}
\details{
The database is stored by default under \code{tools::R_user_dir("tradepolicy")}, or
its location can be set with the environment variable \code{TRADEPOLICY_DIR}.
}
\examples{
\dontrun{ tp_download() }
}
