% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{tp_fixed_effects}
\alias{tp_fixed_effects}
\title{Extract fixed effects from regression object}
\usage{
tp_fixed_effects(fit)
}
\arguments{
\item{fit}{regression object}
}
\value{
a tibble
}
\description{
Takes an lm/glm object and extracts the fixed effects estimated coefficients.
This function was created to be used with \code{left_join()} and \code{predict()} as it
pastes the effects and allows to create a column with the predicted output.
}
