% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{tp_pane}
\alias{tp_pane}
\title{Open AGTPA database connection pane in RStudio}
\usage{
tp_pane()
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) tp_pane()
}
