% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{tp_summary_app_1}
\alias{tp_summary_app_1}
\title{Application 1 (Traditional Gravity Estimates) Reporting Style}
\usage{
tp_summary_app_1(
  formula,
  data,
  method = "ppml",
  pair = "pair_id",
  etfe = "exp_year",
  itfe = "imp_year"
)
}
\arguments{
\item{formula}{Formula for the model}

\item{data}{Tibble or data.frame}

\item{method}{Regression method, which can be "ols" or "ppml" (default)}

\item{pair}{Inter-national fixed effects column (defaults to "pair_id")}

\item{etfe}{Exporter time fixed effects column (defaults to "exp_year")}

\item{itfe}{Importer time fixed effects column (defaults to "imp_year")}
}
\value{
A list
}
\description{
Computes clustered standard errors, tests on coefficients with
clustered standard errors and obtains RESET test p-value.
}
\examples{
# See the ebook

}
