% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{tp_summary_app_2}
\alias{tp_summary_app_2}
\title{Application 2 (The "Distance Puzzle" Resolved) Reporting Style}
\usage{
tp_summary_app_2(
  formula,
  data,
  method = "ppml",
  pair = "pair_id",
  etfe = "exp_year",
  itfe = "imp_year",
  dist = "log_dist",
  intr = "log_dist_intra",
  csfe = "intra_pair"
)
}
\arguments{
\item{formula}{Formula for the model}

\item{data}{Tibble or data.frame}

\item{method}{Regression method (lm or glm)}

\item{pair}{Inter-national fixed effects column (defaults to "pair_id")}

\item{etfe}{Exporter time fixed effects column (defaults to "exp_year")}

\item{itfe}{Importer time fixed effects column (defaults to "imp_year")}

\item{dist}{Distance column (defaults to "log_dist")}

\item{intr}{Intra-national distance column (defaults to "log_dist_intra")}

\item{csfe}{Country-specific fixed effects (defaults to "intra_pair")}
}
\value{
A list
}
\description{
Computes clustered standard errors, tests on coefficients with
clustered standard errors and uses the delta method to obtain changes in
time-based distance estimated coefficients.
}
