% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{tp_table}
\alias{tp_table}
\title{AGTPA applications data}
\usage{
tp_table(tabla)
}
\arguments{
\item{table}{A string indicating the table to extract}
}
\value{
A tibble
}
\description{
Returns a remote database table with the data required to replicate the
exercises from the book. For pre-filtered data please use
\code{\link[=tp_database]{tp_database()}}.
}
\examples{
\dontrun{ tp_table("ch1_application1") }
}
