% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tradepolicy_data}
\alias{tradepolicy_data}
\title{Yotov applications data}
\usage{
tradepolicy_data(table)
}
\arguments{
\item{table}{A string indicating the table to extract}
}
\value{
A \strong{dplyr} tibble (\code{\link[dplyr:tbl]{dplyr::tbl()}})
}
\description{
Returns a remote database table with the data required to replicate the
exercises from the book.
}
\examples{
if (tradepolicy_status()) {
  tradepolicy_data("ch1_application1")
}
}
