% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tradepolicy_db}
\alias{tradepolicy_db}
\title{The local Yotov database}
\usage{
tradepolicy_db(dbdir = tradepolicy_path())
}
\arguments{
\item{dbdir}{The location of the database on disk. Defaults to
\code{tradepolicy} under \code{tools::R_user_dir("tradepolicy")}, or the path specified
in the environment variable \code{TRADEPOLICY_DB_DIR}.}
}
\description{
Returns a connection to the local tradepolicy database. This is a DBI-compliant
duckdb database connection. When using \strong{dplyr}-based
workflows, one typically accesses tables with \code{\link[=tradepolicy_data]{tradepolicy_data()}}, but this
function lets the user interact with the database directly via SQL.
}
\examples{
if (tradepolicy_status()) {
  DBI::dbListTables(tradepolicy_db())

  ch1_application1 <- DBI::dbReadTable(tradepolicy_db(), "ch1_application1")

  DBI::dbGetQuery(
    tradepolicy_db(),
    "SELECT * FROM ch1_application1"
  )
}
}
