% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{tradepolicy_db_delete}
\alias{tradepolicy_db_delete}
\title{Remove the local Yotov database}
\usage{
tradepolicy_db_delete()
}
\description{
Deletes all tables from the local database.
}
\examples{
\donttest{
\dontrun{
tradepolicy_db_delete()
}
}
}
