% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tradepolicy_db_disconnect}
\alias{tradepolicy_db_disconnect}
\title{Disconnect from the Yotov database}
\usage{
tradepolicy_db_disconnect()
}
\description{
A utility function for disconnecting from the database.
}
\examples{
tradepolicy_db_disconnect()
}
