% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{tradepolicy_pane}
\alias{tradepolicy_pane}
\title{Open Yotov database connection pane in RStudio}
\usage{
tradepolicy_pane()
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) tradepolicy_pane()
}
