% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tradepolicy_status}
\alias{tradepolicy_status}
\title{Get the status of the current local Yotov database}
\usage{
tradepolicy_status(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print a status message}
}
\value{
TRUE if the database exists, FALSE if it is not detected. (invisible)
}
\description{
Get the status of the current local Yotov database. It displays informative message
about how to create the local database if it can't be found or it is corrupt.
}
\examples{
tradepolicy_status()
}
