% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{yotov_data}
\alias{yotov_data}
\title{Yotov applications data}
\usage{
yotov_data(table)
}
\arguments{
\item{table}{A string indicating the table to extract}
}
\value{
A \strong{dplyr} tibble (\code{\link[dplyr:tbl]{dplyr::tbl()}})
}
\description{
Returns a remote database table with the data required to replicate the
exercises from the book.
}
\examples{
if (yotov_status()) {
  # See the number of yotov shipment records per year
  yotov_data("ch1_application1") \%>\%
    filter(exporter == "ARG")
}
}
