% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{yotov_db_delete}
\alias{yotov_db_delete}
\title{Remove the local Yotov database}
\usage{
yotov_db_delete()
}
\description{
Deletes all tables from the local database
}
\examples{
\donttest{
\dontrun{
yotov_db_delete()
}
}
}
