% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{yotov_db_disconnect}
\alias{yotov_db_disconnect}
\title{Disconnect from the Yotov database}
\usage{
yotov_db_disconnect()
}
\description{
A utility function for disconnecting from the database.
}
\examples{
yotov_db_disconnect()
}
