% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{yotov_fixed_effects}
\alias{yotov_fixed_effects}
\title{GLM Regression With Robust Clustered Standard Errors}
\usage{
yotov_fixed_effects(fit)
}
\arguments{
\item{fit}{A regression object}
}
\description{
Computes clustered standard errors, tests on coefficients with
clustered standard errors and delta method for percent change in log
}
