% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection-pane.R
\name{yotov_pane}
\alias{yotov_pane}
\title{Open Yotov database connection pane in RStudio}
\usage{
yotov_pane()
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
if (!is.null(getOption("connectionObserver"))) yotov_pane()
}
