% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{yotov_robust_glm}
\alias{yotov_robust_glm}
\title{GLM Regression With Robust Clustered Standard Errors}
\usage{
yotov_robust_glm(formula, data)
}
\arguments{
\item{formula}{A formula for the model}

\item{data}{A tibble or data.frame}
}
\description{
Computes clustered standard errors, tests on coefficients with
clustered standard errors and delta method for percent change in log
}
