% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{yotov_status}
\alias{yotov_status}
\title{Get the status of the current local Yotov database}
\usage{
yotov_status(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print a status message}
}
\value{
TRUE if the database exists, FALSE if it is not detected. (invisible)
}
\description{
Get the status of the current local Yotov database
}
\examples{
yotov_status()
}
