% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{tp_database}
\alias{tp_database}
\title{The local Yotov database}
\usage{
tp_database(dir = tp_path())
}
\arguments{
\item{dir}{The location of the database on disk. Defaults to
\code{tradepolicy} under \code{tools::R_user_dir("tradepolicy")}, or the path specified
in the environment variable \code{TRADEPOLICY_DIR}.}
}
\description{
Returns a connection to the local tradepolicy database. This is a DBI-compliant
duckdb database connection. When using \strong{dplyr}-based
workflows, one typically accesses tables with \code{\link[=tp_data]{tp_data()}}, but this
function lets the user interact with the database directly via SQL.
}
\examples{
if (tp_status()) {
  DBI::dbListTables(tp_database())

  ch1_application1 <- DBI::dbReadTable(tp_database(), "ch1_application1")

  DBI::dbGetQuery(
    tp_database(),
    "SELECT * FROM ch1_application1"
  )
}
}
