% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradepolicy-package.R
\docType{package}
\name{tradepolicy-package}
\alias{tradepolicy}
\alias{tradepolicy-package}
\title{tradepolicy: Replication of 'An Advanced Guide To Trade Policy Analysis'}
\description{
Datasets from An Advanced Guide to Trade Policy Analysis
    (Year: 2016, ISBN: 978-92-870-4367-2) by Yotov, et al. and functions to 
    report regression summaries with clustered robust standard errors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.tiid.org/R_structural_gravity/}
  \item Report bugs at \url{https://github.com/pachamaltese/tradepolicy/issues/}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas \email{mavargas11@uc.cl} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Other contributors:
\itemize{
  \item Alexey Kravchenko [thesis advisor]
  \item Constanza Prado [thesis advisor]
  \item Yoto Yotov [thesis advisor]
  \item  The United Nations [data contributor, copyright holder, funder]
}

}
\keyword{internal}
