% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust-summary.R
\name{yotov_clustered_summary}
\alias{yotov_clustered_summary}
\title{Work in progress function to Report Clusters Standard Errors in Summaries}
\usage{
yotov_clustered_summary(model, data, robust = FALSE, cluster = "pair_id")
}
\arguments{
\item{model}{explain}

\item{data}{explain}

\item{robust}{explain}

\item{cluster}{explain}
}
\description{
Work in progress function to Report Clusters Standard Errors in Summaries
}
