% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{yotov_fixed_effects}
\alias{yotov_fixed_effects}
\title{Extract fixed effects from regression object}
\usage{
yotov_fixed_effects(fit)
}
\arguments{
\item{fit}{A regression object}
}
\description{
Complete description
}
