% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1.R
\name{yotov_model_summary2}
\alias{yotov_model_summary2}
\title{The "Distance Puzzle" Resolved Reporting Style}
\usage{
yotov_model_summary2(formula, data, method)
}
\arguments{
\item{formula}{A formula for the model}

\item{data}{A tibble or data.frame}

\item{method}{Regression method (lm or glm)}
}
\description{
Computes clustered standard errors, tests on coefficients with
clustered standard errors and delta method for percent change in log.
}
