% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter2.R
\name{yotov_robust_glm}
\alias{yotov_robust_glm}
\title{GLM Regression With Robust Clustered Standard Errors}
\usage{
yotov_robust_glm(formula, data)
}
\arguments{
\item{formula}{A formula for the model}

\item{data}{A tibble or data.frame}
}
\description{
Fits a regression with robust clustered standard errors. This uses a quasi-poisson
family and returns the estimated coefficients after computing a clustered
variance-covariance matrix.
}
